+++ CHANGELOG.md
@@ -27,6 +27,9 @@ of requirements for an API to count as public.
   and `BaseThrottleSyncBackend`, #942
 - *Breaking*: Renamed `DjangoCache` into `DjangoSyncCache`,
   added `DjangoAsyncCache`, #942
+- *Breaking*: Removed `BaseThrottleAlgorithm.record` method,
+  now `BaseThrottleAlgorithm.access` must also record accesses.
+  This will help to make throttling more atomic, #942
 - Added `RefreshTokenSyncController` and `RefreshTokenAsyncController`
   to issue new access/refresh token pairs from a valid refresh token, #907
 - Added `validate_negotiation` metadata flag, so we can explicitly validate,